+++ CHANGES.md
@@ -3,6 +3,9 @@
 ## develop
 
 - [CHANGE] cowboy 2.0 で onresponse オプションが無くなったため削除する
+- [ADD] interceptor オプションを追加する
+  - dispatch 前に interceptor モジュールを呼び出し、JSON の変更、およびリクエストの中断を
+    可能にする
 
 ## 2021.3